bug TMDM-2574 fixed: fix delete action error
git-svn-id: http://talendforge.org/svn/tom/trunk@69494 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.mdm.webapp.browserecords/src/main/java/org/talend/mdm/webapp/browserecords/server/actions/BrowseRecordsAction.java
@@ -46,6 +46,8 @@ import javax.xml.xpath.XPathExpressionException;
 
 import org.apache.commons.lang.StringUtils;
 import org.apache.log4j.Logger;
+import org.talend.mdm.commmon.util.core.EDBType;
+import org.talend.mdm.commmon.util.core.MDMConfiguration;
 import org.talend.mdm.commmon.util.datamodel.management.BusinessConcept;
 import org.talend.mdm.commmon.util.datamodel.management.ReusableType;
 import org.talend.mdm.commmon.util.webapp.XSystemObjects;
@@ -56,7 +58,6 @@ import org.talend.mdm.webapp.base.client.model.ItemBaseModel;
 import org.talend.mdm.webapp.base.client.model.ItemBasePageLoadResult;
 import org.talend.mdm.webapp.base.client.model.SubTypeBean;
 import org.talend.mdm.webapp.base.server.BaseConfiguration;
-import org.talend.mdm.webapp.base.server.ForeignKeyHelper;
 import org.talend.mdm.webapp.base.server.util.CommonUtil;
 import org.talend.mdm.webapp.base.shared.TypeModel;
 import org.talend.mdm.webapp.browserecords.client.BrowseRecordsService;
@@ -107,6 +108,8 @@ import com.amalto.webapp.core.util.XmlUtil;
 import com.amalto.webapp.core.util.XtentisWebappException;
 import com.amalto.webapp.util.webservices.WSBoolean;
 import com.amalto.webapp.util.webservices.WSByteArray;
+import com.amalto.webapp.util.webservices.WSCount;
+import com.amalto.webapp.util.webservices.WSCountItemsByCustomFKFilters;
 import com.amalto.webapp.util.webservices.WSDataClusterPK;
 import com.amalto.webapp.util.webservices.WSDataModelPK;
 import com.amalto.webapp.util.webservices.WSDeleteItem;
@@ -117,6 +120,7 @@ import com.amalto.webapp.util.webservices.WSExistsItem;
 import com.amalto.webapp.util.webservices.WSGetBusinessConcepts;
 import com.amalto.webapp.util.webservices.WSGetDataModel;
 import com.amalto.webapp.util.webservices.WSGetItem;
+import com.amalto.webapp.util.webservices.WSGetItemsByCustomFKFilters;
 import com.amalto.webapp.util.webservices.WSGetTransformer;
 import com.amalto.webapp.util.webservices.WSGetTransformerPKs;
 import com.amalto.webapp.util.webservices.WSGetView;
@@ -187,15 +191,25 @@ public class BrowseRecordsAction implements BrowseRecordsService {
                 }
             }
 
-            if (outputErrorMessage == null || "info".equals(errorCode)) { //$NON-NLS-1$
-                WSItemPK wsItem = CommonUtil.getPort().deleteItem(
-                        new WSDeleteItem(new WSItemPK(new WSDataClusterPK(dataClusterPK), concept, ids), override));
-                if (wsItem != null) {
-                    pushUpdateReport(ids, concept, "PHYSICAL_DELETE", true); //$NON-NLS-1$
-                    if (message == null || message.length() == 0)
-                        message = MESSAGES.getMessage("delete_record_success"); //$NON-NLS-1$
+            if (outputErrorMessage == null || "info".equals(errorCode)) { //$NON-NLS-1$                
+                if (ids != null && !item.isReadOnly()) {
+                    WSItemPK wsItem = CommonUtil.getPort().deleteItem(
+                            new WSDeleteItem(new WSItemPK(new WSDataClusterPK(dataClusterPK), concept, ids), override));
+                    if (wsItem != null)
+                        pushUpdateReport(ids, concept, "PHYSICAL_DELETE", true); //$NON-NLS-1$
+                    else
+                        throw new ServiceException(MESSAGES.getMessage("delete_record_failure")); //$NON-NLS-1$
+
+                    if (outputErrorMessage == null)
+                        message = message == null ? "" : message; //$NON-NLS-1$
+                    else if (message == null || message.length() == 0)
+                        message = MESSAGES.getMessage("delete.process.validation.success"); //$NON-NLS-1$
                 } else {
-                    throw new ServiceException(MESSAGES.getMessage("delete_record_failure")); //$NON-NLS-1$
+                    if (outputErrorMessage == null)
+                        message = message == null ? "" : message; //$NON-NLS-1$
+                    else if (message == null || message.length() == 0)
+                        message = MESSAGES.getMessage("delete.process.validation.success"); //$NON-NLS-1$
+                    return MESSAGES.getMessage("delete_item_record_successNoupdate", message); //$NON-NLS-1$
                 }
             } else {
                 // Anything but 0 is unsuccessful
@@ -203,13 +217,14 @@ public class BrowseRecordsAction implements BrowseRecordsService {
                     message = MESSAGES.getMessage("delete_process_validation_failure"); //$NON-NLS-1$
                 throw new ServiceException(message);
             }
+
             return message;
         } catch (ServiceException e) {
             LOG.error(e.getMessage(), e);
             throw e;
         } catch (Exception e) {
             LOG.error(e.getMessage(), e);
-            throw new ServiceException(e.getLocalizedMessage());
+            throw new ServiceException(e.getMessage());
         }
     }
 
@@ -262,8 +277,144 @@ public class BrowseRecordsAction implements BrowseRecordsService {
 
     public ItemBasePageLoadResult<ForeignKeyBean> getForeignKeyList(PagingLoadConfig config, TypeModel model,
             String dataClusterPK, boolean ifFKFilter, String value) throws ServiceException {
+        String xpathForeignKey = model.getForeignkey();
+        // to verify
+        String xpathInfoForeignKey = model.getForeignKeyInfo().toString().replaceAll("\\[", "").replaceAll("\\]", ""); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$ //$NON-NLS-4$
+        // in search panel, the fkFilter is empty
+        String fkFilter = ""; //$NON-NLS-1$
+        if (ifFKFilter)
+            fkFilter = model.getFkFilter();
+
+        if (xpathForeignKey == null)
+            return null;
+
+        List<ForeignKeyBean> fkBeans = new ArrayList<ForeignKeyBean>();
+        String[] results = null;
+        String count = null;
+
         try {
-            return ForeignKeyHelper.getForeignKeyList(config, model, dataClusterPK, ifFKFilter, value);
+            String initxpathForeignKey = ""; //$NON-NLS-1$
+            initxpathForeignKey = com.amalto.webapp.core.util.Util.getForeignPathFromPath(xpathForeignKey);
+
+            WSWhereCondition whereCondition = com.amalto.webapp.core.util.Util.getConditionFromPath(xpathForeignKey);
+            WSWhereItem whereItem = null;
+            if (whereCondition != null) {
+                whereItem = new WSWhereItem(whereCondition, null, null);
+            }
+
+            // get FK filter
+            WSWhereItem fkFilterWi = com.amalto.webapp.core.util.Util.getConditionFromFKFilter(xpathForeignKey, xpathForeignKey,
+                    fkFilter);
+            if (fkFilterWi != null)
+                whereItem = fkFilterWi;
+            initxpathForeignKey = initxpathForeignKey.split("/")[0]; //$NON-NLS-1$
+
+            xpathInfoForeignKey = xpathInfoForeignKey == null ? "" : xpathInfoForeignKey; //$NON-NLS-1$
+            // foreign key set by business concept
+            if (initxpathForeignKey.split("/").length == 1) { //$NON-NLS-1$
+                String conceptName = initxpathForeignKey;
+                // determine if we have xPath Infos: e.g. labels to display
+                String[] xpathInfos = new String[1];
+                if (!"".equals(xpathInfoForeignKey) && xpathInfoForeignKey != null)//$NON-NLS-1$
+                    xpathInfos = xpathInfoForeignKey.split(","); //$NON-NLS-1$
+                else
+                    xpathInfos[0] = conceptName;
+                value = value == null ? "" : value; //$NON-NLS-1$
+
+                // build query - add a content condition on the pivot if we search for a particular value
+                String filteredConcept = conceptName;
+
+                if (value != null && !"".equals(value.trim()) && !".*".equals(value.trim())) { //$NON-NLS-1$ //$NON-NLS-2$
+                    List<WSWhereItem> condition = new ArrayList<WSWhereItem>();
+                    if (whereItem != null)
+                        condition.add(whereItem);
+                    WSWhereItem wc = null;
+                    String strConcept = conceptName + "/. CONTAINS "; //$NON-NLS-1$
+
+                    if (MDMConfiguration.getDBType().getName().equals(EDBType.QIZX.getName())) {
+                        strConcept = conceptName + "//* CONTAINS "; //$NON-NLS-1$
+                    }
+                    wc = com.amalto.webapp.core.util.Util.buildWhereItem(strConcept + value);
+                    condition.add(wc);
+                    WSWhereAnd and = new WSWhereAnd(condition.toArray(new WSWhereItem[condition.size()]));
+                    WSWhereItem whand = new WSWhereItem(null, and, null);
+                    if (whand != null)
+                        whereItem = whand;
+                }
+
+                // add the xPath Infos Path
+                ArrayList<String> xPaths = new ArrayList<String>();
+                if (model.isRetrieveFKinfos())
+                    // add the xPath Infos Path
+                    for (int i = 0; i < xpathInfos.length; i++) {
+                        xPaths.add(com.amalto.webapp.core.util.Util.getFormatedFKInfo(
+                                xpathInfos[i].replaceFirst(conceptName, filteredConcept), filteredConcept));
+                    }
+                // add the key paths last, since there may be multiple keys
+                xPaths.add(filteredConcept + "/../../i"); //$NON-NLS-1$
+                // order by
+                String orderbyPath = null;
+                if (!MDMConfiguration.getDBType().getName().equals(EDBType.QIZX.getName())) {
+                    if (!"".equals(xpathInfoForeignKey) && xpathInfoForeignKey != null) { //$NON-NLS-1$
+                        orderbyPath = com.amalto.webapp.core.util.Util.getFormatedFKInfo(
+                                xpathInfos[0].replaceFirst(conceptName, filteredConcept), filteredConcept);
+                    }
+                }
+
+                // Run the query
+                if (!com.amalto.webapp.core.util.Util.isCustomFilter(fkFilter)) {
+
+                    results = CommonUtil
+                            .getPort()
+                            .xPathsSearch(
+                                    new WSXPathsSearch(new WSDataClusterPK(dataClusterPK), null, new WSStringArray(xPaths
+                                            .toArray(new String[xPaths.size()])), whereItem, -1, config.getOffset(), config
+                                            .getLimit(), orderbyPath, null)).getStrings();
+                    count = CommonUtil.getPort()
+                            .count(new WSCount(new WSDataClusterPK(dataClusterPK), conceptName, whereItem, -1)).getValue();
+
+                } else {
+
+                    String injectedXpath = com.amalto.webapp.core.util.Util.getInjectedXpath(fkFilter);
+                    results = CommonUtil
+                            .getPort()
+                            .getItemsByCustomFKFilters(
+                                    new WSGetItemsByCustomFKFilters(new WSDataClusterPK(dataClusterPK), conceptName,
+                                            new WSStringArray(xPaths.toArray(new String[xPaths.size()])), injectedXpath, config
+                                                    .getOffset(), config.getLimit(), orderbyPath, null)).getStrings();
+
+                    count = CommonUtil
+                            .getPort()
+                            .countItemsByCustomFKFilters(
+                                    new WSCountItemsByCustomFKFilters(new WSDataClusterPK(dataClusterPK), conceptName,
+                                            injectedXpath)).getValue();
+                }
+            }
+
+            if (results != null) {
+                for (String result : results) {
+                    ForeignKeyBean bean = new ForeignKeyBean();
+                    String id = ""; //$NON-NLS-1$
+                    NodeList nodes = Util.getNodeList(Util.parse(result), "//i"); //$NON-NLS-1$
+                    if (nodes != null) {
+                        for (int i = 0; i < nodes.getLength(); i++) {
+                            if (nodes.item(i) instanceof Element)
+                                id += "[" + (nodes.item(i).getTextContent() == null ? "" : nodes.item(i).getTextContent()) + "]"; //$NON-NLS-1$  //$NON-NLS-2$  //$NON-NLS-3$
+                        }
+                    }
+                    bean.setId(id);
+                    if (result != null) {
+                        Element root = Util.parse(result).getDocumentElement();
+                        if (root.getNodeName().equals("result"))//$NON-NLS-1$
+                            initFKBean(root, bean);
+                        else
+                            bean.set(root.getNodeName(), root.getTextContent().trim());
+                    }
+                    fkBeans.add(bean);
+                }
+            }
+
+            return new ItemBasePageLoadResult<ForeignKeyBean>(fkBeans, config.getOffset(), Integer.valueOf(count));
         } catch (Exception e) {
             LOG.error(e.getMessage(), e);
             throw new ServiceException(e.getLocalizedMessage());
@@ -287,7 +438,7 @@ public class BrowseRecordsAction implements BrowseRecordsService {
         }
         return xPathList;
     }
-
+    
     private String getPKInfos(List<String> xPathList) {
         StringBuilder gettedValue = new StringBuilder();
         for (String pkInfo : xPathList) {
@@ -810,9 +961,9 @@ public class BrowseRecordsAction implements BrowseRecordsService {
                     err = e.getLocalizedMessage();
                 }
                 // add feature TMDM-2327 SAXException:cvc-complex-type.2.4.b message transform
-                if (e.getLocalizedMessage().indexOf("cvc-complex-type.2.4.b") != -1) { 
-                    err = "Unable to save item,before saving the '" + item.getConcept() 
-                            + "' item,please fill the required field's contents";
+                if (e.getLocalizedMessage().indexOf("cvc-complex-type.2.4.b") != -1) { //$NON-NLS-1$
+                    err = "Unable to save item,before saving the '" + item.getConcept() //$NON-NLS-1$
+                            + "' item,please fill the required field's contents"; //$NON-NLS-1$
                 }
                 serviceException = new ServiceException(err);
             }
@@ -1487,21 +1638,18 @@ public class BrowseRecordsAction implements BrowseRecordsService {
         } catch (Exception e) {
             // TODO UGLY!!!! to be refactored
             ServiceException serviceException;
-            if (e.getLocalizedMessage().indexOf("routing failed:") == 0) {
-                String saveSUCCE = "Save item '" + concept + "."
-                        + com.amalto.webapp.core.util.Util.joinStrings(convertIds(ids), ".") + "' successfully, But "
-                        + e.getLocalizedMessage();
-                serviceException = new ServiceException(saveSUCCE);
+            if (e.getMessage().indexOf("routing failed:") == 0) { //$NON-NLS-1$
+                serviceException = new ServiceException(MESSAGES.getMessage("save_successEx", concept + "." //$NON-NLS-1$ //$NON-NLS-2$
+                        + com.amalto.webapp.core.util.Util.joinStrings(convertIds(ids), "."), e.getMessage())); //$NON-NLS-1$
             } else {
-                String err = "Unable to save item '" + concept + "."
-                        + com.amalto.webapp.core.util.Util.joinStrings(convertIds(ids), ".") + "'" + e.getLocalizedMessage();
-                if (e.getLocalizedMessage().indexOf("ERROR_3:") == 0) {
-                    err = e.getLocalizedMessage();
+                String err = MESSAGES.getMessage("save_fail", concept + "." //$NON-NLS-1$ //$NON-NLS-2$
+                        + com.amalto.webapp.core.util.Util.joinStrings(convertIds(ids), ".") + e.getMessage()); //$NON-NLS-1$
+                if (e.getMessage().indexOf("ERROR_3:") == 0) { //$NON-NLS-1$
+                    err = e.getMessage();
                 }
                 // add feature TMDM-2327 SAXException:cvc-complex-type.2.4.b message transform
-                if (e.getLocalizedMessage().indexOf("cvc-complex-type.2.4.b") != -1) {
-                    err = "Unable to save item,before saving the '" + concept
-                            + "' item,please fill the required field's contents"; 
+                if (e.getMessage().indexOf("cvc-complex-type.2.4.b") != -1) { //$NON-NLS-1$
+                    err = MESSAGES.getMessage("save_failEx", concept); //$NON-NLS-1$
                 }
                 serviceException = new ServiceException(err);
             }
